### 2022-10-29 #4891 @editorialbot
[PRE REVIEW]: PhaseTypeR: an R package for phase-type distributions in population genetics
**Submitting author:** @rivasiker (<a href="http://orcid.org/0000-0002-0515-0628">Iker Rivas-González</a>)
**Repository:** https://github.com/rivasiker/PhaseTypeR
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.4
