+++ pkgdown.yml
@@ -5,5 +5,5 @@ articles:
   PhaseTypeR: PhaseTypeR.html
   a2_pop_gen_iker: a2_pop_gen_iker.html
   a3_phasetypeR_SFS: a3_phasetypeR_SFS.html
-last_built: 2022-09-23T17:38Z
+last_built: 2022-11-01T13:42Z
 
