+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={81},
   year={2017},
   publisher={Springer-Verlag},
-  isbn={978-1-4939-8377-3}
+  isbn={978-1-4939-8377-3},
+  DOI={10.1007/978-1-4939-7049-0}
 }
 
 @article{hobolth2019phase,
@@ -16,7 +17,7 @@
   pages={16--32},
   year={2019},
   publisher={Elsevier},
-  DOI={https://doi.org/10.1016/j.tpb.2019.02.001}
+  DOI={10.1016/j.tpb.2019.02.001}
 }
 
 @article{dutang2008actuar,
@@ -37,7 +38,8 @@
   number={1},
   pages={44--77},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  DOI={10.1111/sjos.12505}
 }
 
 @article{albrecher_bladt_2019, 
@@ -60,7 +62,7 @@ journal={Journal of the Operations Research Society of Japan},
 author={Okamura, H and Dohi, T}, 
 year={2016}, 
 pages={72-109},
-DOI={https://doi.org/10.15807/jorsj.59.72}
+DOI={10.15807/jorsj.59.72}
 }
 
 @InProceedings{okamura_dohi_2015,
@@ -73,7 +75,8 @@ booktitle="Quantitative Evaluation of Systems",
 year="2015",
 publisher="Springer International Publishing",
 pages="105--112",
-isbn={978-3-319-22263-9}
+isbn={978-3-319-22263-9},
+DOI={10.1007/978-3-319-22264-6_7}
 }
 
 @Manual{okamura2015mapfit,
@@ -93,7 +96,7 @@ isbn={978-3-319-22263-9}
   pages={1--28},
   year={2021},
   publisher={Springer},
-  DOI={https://doi.org/10.1007/s00285-021-01689-w}
+  DOI={10.1007/s00285-021-01689-w}
 }
 
 @phdthesis{navarro2019discrete,
@@ -113,7 +116,8 @@ note = {ISSN 0909-3192}
   author={Durrett, Richard},
   volume={2},
   year={2008},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/978-0-387-78168-6}
 }
 
 @book{wakeley2009coalescent,
@@ -125,6 +129,7 @@ note = {ISSN 0909-3192}
         publisher = {W. H. Freeman},
         author = {Wakeley, , John},
         year = {2009},
+        DOI={10.1093/schbul/syp004}
 }
 
 @article{wilton2015smc,
@@ -136,7 +141,7 @@ note = {ISSN 0909-3192}
   pages={343--355},
   year={2015},
   publisher={Oxford University Press},
-  DOI={https://doi.org/10.1534/genetics.114.173898}
+  DOI={10.1534/genetics.114.173898}
 }
 
 @article{simonsen1997markov,
@@ -148,7 +153,7 @@ note = {ISSN 0909-3192}
   pages={43--59},
   year={1997},
   publisher={Elsevier},
-  DOI={https://doi.org/10.1006/tpbi.1997.1307}
+  DOI={10.1006/tpbi.1997.1307}
 }
 
 @article{kingman1982coalescent,
@@ -160,7 +165,7 @@ note = {ISSN 0909-3192}
   pages={235--248},
   year={1982},
   publisher={Elsevier},
-  DOI={https://doi.org/10.1016/0304-4149(82)90011-4}
+  DOI={10.1016/0304-4149(82)90011-4}
 }
 
 @Manual{Rcitation,
@@ -180,7 +185,8 @@ note = {ISSN 0909-3192}
   number={2},
   pages={172--197},
[…]
