### 2022-12-14 #4891 @majensen
Hi @ethanbass -- would you be able to review [this work](#4891) in [JOSS](https://joss.theoj.org)? We could use your expertise in this area. Thanks for your consideration! --@majensen
