### 2022-12-18 #4891 @ethanbass
> Hi @ethanbass -- would you be able to review [this work](#4891) in [JOSS](https://joss.theoj.org)? We could use your expertise in this area. Thanks for your consideration! --@majensen

Thank you for thinking of me. I feel I may not be so qualified to review this submission. I am not really a popgen person and I feel that some of this may be over my head.
