### 2022-12-19 #4891 @dwinter
Hi @majensen . Sincere apologies, this seems like a cool problem and program but my availability to review it now is very limited between work and finally taking an extended  break from work over the Southern Summer. I don't think I would be able to start a review until February at the very earliest I'm sorry.
