### 2022-12-20 #4891 @majensen
Hi @tkchafin - would you be willing and available to review [this work](#4891) for [JOSS](https://joss.theoj.org)? I could use your expertise in population genetics. Thanks for your consideration --@majensen
