### 2022-12-20 #4891 @tkchafin
Yes I can review, but with the holidays and personal leave I am taking it would take longer than usual. Sent via the Samsung Galaxy S10, an AT&T 5G Evolution capable smartphone
-------- Original message --------From: Mark Jensen ***@***.***> Date: 20/12/2022  2:37 am  (GMT+00:00) To: openjournals/joss-reviews ***@***.***> Cc: Tyler Chafin ***@***.***>, Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: PhaseTypeR: an R
  package for phase-type distributions in population genetics (Issue #4891)
Hi @tkchafin - would you be willing and available to review this work for JOSS? I could use your expertise in population genetics. Thanks for your consideration ***@***.***
