### 2023-01-03 #4891 @majensen
Hi @cboettig - would you be willing to review [this work](4891) for [JOSS](https://joss.theoj.org)? I believe you might fit in the intersection of the technical and biological aspects of this paper. Thanks for considering! --@majensen
