### 2023-02-05 rivasiker/PhaseTypeR#9 @tkchafin
[JOSS Review] Missing dependencies for vignette3 (SFS)
Vignette 3 missing definition of dependencies (partitions and expm). Should at minimum be added as the first instruction to install these. You may also want to add a subheading to the # Install section of the Readme.Md listing the required dependencies for running PhaseTypeR (although I realize these are handled for you when e.g. installing from CRAN)
