### 2023-02-05 rivasiker/PhaseTypeR#10 @tkchafin
[JOSS Review] Add issues/ help section to Readme
Per JOSS review, there should be some instruction provided on how users should reach out to report issues/ get help. Most likely just linking to the GitHub Issues page, but should mention the recommended procedure in a section of the Readme
