### 2023-02-05 rivasiker/PhaseTypeR#11 @tkchafin
[JOSS Review] Minor MS suggestions 
Very minor suggestions for the manuscript, none of which I consider essential by any means
- L32-34: Would read better with removing the redundant “they” after each comma (L33 “they do not” -> “do not” and L34 “and they are” -> “and are”)
- L35: Maybe change to “Here we present an alternative, PhaseTypeR which is particularly well suited for population genetics, in that it solves these limitations with easy-to-use R functions”, or similar, since the purpose is to presumably highlight that this package is meant to address the issues listed in previous sentence.
- Because the Overview section (start Line 40) is a bit technical, it might make sense to add a few short sentences at the beginning (before paragraph starting L41 with “Table 1” giving a higher-level description of the functions/ distribution types. Something like “PhaseTypeR contains functions for doing X, Y and Z from univariate continuous, univariate discrete, or multivariate phase-type distributions etc etc”
- L180: Rogue quotation mark after “SMC”?
