### 2023-02-05 #5054 @tkchafin
I've completed my review -- provided some very minor suggestions and flagged a few missing bits to meet the community guidelines as issues [#8](https://github.com/rivasiker/PhaseTypeR/issues/8), [#9](https://github.com/rivasiker/PhaseTypeR/issues/9), [#10](https://github.com/rivasiker/PhaseTypeR/issues/10), and  [#11](https://github.com/rivasiker/PhaseTypeR/issues/11).

No issues were encountered installing or running any of the vignettes, other than some undocumented dependencies which were flagged in the aforementioned issues. Once the contributing, issue/ bug report, and dependencies have been documented I'll tick the last box on the checklist. Thanks and congrats to the authors on a well put together R package.
