### 2023-02-06 rivasiker/PhaseTypeR#8 @rivasiker
We thank the reviewer for pointing this out. We have included the two files in [this commit](https://github.com/rivasiker/PhaseTypeR/commit/d1e04d3bf9cd3be366726452b514c0f5b775cf05), by running `usethis::use_tidy_contributing()` and `usethis::use_tidy_coc()`, and modifying the files accordingly.
