+++ .github/CONTRIBUTING.md
@@ -39,5 +39,5 @@ Please, contact the PhaseTypeR team. We'll be happy to discuss your thoughts!
 ### Code of Conduct
 
 Please note that the PhaseTypeR project is released with a
-[Contributor Code of Conduct](.github/CODE_OF_CONDUCT.md). By contributing to this
+[Contributor Code of Conduct](CODE_OF_CONDUCT.md). By contributing to this
 project you agree to abide by its terms.
