### 2023-02-06 rivasiker/PhaseTypeR#9 @rivasiker
We thank the reviewer for pointing this out. We have included instructions for installing the dependencies in vignette 3 and in the README.md file through [this commit](https://github.com/rivasiker/PhaseTypeR/commit/aac30b304ded7eb9d186406df149d08ff6c8a2e6).
