### 2023-02-06 rivasiker/PhaseTypeR#10 @rivasiker
Thank you for pointing this out. We have added a brief description on how to report a bug in the README.md file, together with a link to the CONTRIBUTING.md file, which contains a more detail description on how to file an issue. Changes to the README.md file can be linked to [this commit](https://github.com/rivasiker/PhaseTypeR/commit/74bd698f9c55179a86936153faec239a8d2aea7d).
