### 2023-02-06 rivasiker/PhaseTypeR#11 @rivasiker
Thank you for your suggestions. We have included the first three points in the latest version of our paper (see [this commit](https://github.com/rivasiker/PhaseTypeR/commit/05d77abc9096a6cdb673bbb7f48f19939428837c)). As for the last point, the name of the method is indeed SMC' (with a single quotation mark, see [the corresponding article](https://doi.org/10.1534/genetics.114.173898)).
