### 2023-02-06 d1e04d3
Add CODE_OF_CONDUCT.md and CONTRIBUTING.md
