+++ README.Rmd
@@ -17,7 +17,7 @@ knitr::opts_chunk$set(
 <!-- badges: start -->
 [![R-CMD-check](https://github.com/rivasiker/PhaseTypeR/actions/workflows/check-standard.yaml/badge.svg)](https://github.com/rivasiker/PhaseTypeR/actions/workflows/check-standard.yaml)
 [![CRAN](https://www.r-pkg.org/badges/version/PhaseTypeR)](https://CRAN.R-project.org/package=PhaseTypeR)
-[![devel Version](https://img.shields.io/badge/devel-1.0.2.9000-brightgreen.svg)](https://github.com/rivasiker/PhaseTypeR)
+[![devel Version](https://img.shields.io/badge/devel-1.0.5.9000-brightgreen.svg)](https://github.com/rivasiker/PhaseTypeR)
 [![downloads](https://cranlogs.r-pkg.org/badges/grand-total/PhaseTypeR)](https://CRAN.R-project.org/package=PhaseTypeR)
 [![Codecov test coverage](https://codecov.io/gh/rivasiker/PhaseTypeR/branch/master/graph/badge.svg)](https://app.codecov.io/gh/rivasiker/PhaseTypeR?branch=master)
 <!-- badges: end -->
@@ -37,6 +37,13 @@ If you install devtools in your R environment with `install.packages("devtools")
 
 ``` r
 devtools::install_github("rivasiker/PhaseTypeR")
+```
+
+These installations should handle the missing dependencies, but you can manually install them by running the following:
+
+``` r
+install.packages("expm")
+install.packages("igraph")
 ```
 
 ## Basic example
