+++ README.Rmd
@@ -97,8 +97,6 @@ set.seed(0)
 rPH(6, ph)
 ```
 
-
-
 ## Learn more
 
 You can check out the full functionality of `PhaseTypeR` and its application to 
@@ -108,3 +106,6 @@ population genetics in the following guides:
   - [Using PhaseTypeR for population genetics](https://rivasiker.github.io/PhaseTypeR/articles/a2_pop_gen_iker.html)
   - [Using PhaseTypeR for the site frequency spectrum](https://rivasiker.github.io/PhaseTypeR/articles/a3_phasetypeR_SFS.html)
   
+## Report an issue
+
+If you have found a bug, you can file an issue at <https://github.com/rivasiker/PhaseTypeR/issues>. Please, make sure to read the [contribution guidelines](.github/CONTRIBUTING.md) and the [code of conduct](.github/CODE_OF_CONDUCT.md) before making any changes. 
