+++ README.md
@@ -130,3 +130,10 @@ application to population genetics in the following guides:
   genetics](https://rivasiker.github.io/PhaseTypeR/articles/a2_pop_gen_iker.html)
 - [Using PhaseTypeR for the site frequency
   spectrum](https://rivasiker.github.io/PhaseTypeR/articles/a3_phasetypeR_SFS.html)
+
+## Report an issue
+
+If you have found a bug, you can file an issue at
+<https://github.com/rivasiker/PhaseTypeR/issues>. Please, make sure to
+read the [contribution guidelines](.github/CONTRIBUTING.md) and the
+[code of conduct](.github/CODE_OF_CONDUCT.md) before making any changes.
