### 2023-02-09 #5054 @nhejazi
I've just completed my checklist. I encountered only the issues already reported by the other reviewer, which appear to have now been fixed on the `joss-paper` branch. The package is very well put together, and the paper is also succinct and informative -- congratulations to the authors on such a nice product.
