### 2023-02-18 #5054 @majensen
@rivasiker all I can say is, I wish I'd had phase-type distributions and this nice package back in the late 90s! The paper reads very well, and I don't see any issues. Before recommending for publication, can you create a public archive of the codebase using Zenodo, FigShare, or similar? (see https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729 for a short tutorial if needed).
Please let me know in this thread the DOI you get.
