### 2023-02-20 #5054 @rivasiker
@majensen, thank you very much for your kind words! I have just created a release of the package in Zenodo with DOI: 10.5281/zenodo.7656038 (https://doi.org/10.5281/zenodo.7656038). Please, let me know if there is something else I should do.
