### 2023-02-20 #5054 @rivasiker
@majensen Looking at the final proof, I have noticed that Table 2 spans two pages. I think it could span just one by adjusting the width of the table columns. Additionally, the two warnings that appeared during the generation of the final proof could also be easily fixed. Should I go ahead and fix these two details? After doing so, should I re-generate the proof by running `@editorialbot generate pdf`?
