+++ paper/paper.md
@@ -97,14 +97,14 @@ and `r` is the reward vector.\label{tab:tab1}
 
 
 | Quantity                  | Formula                                                                                | Function                    |
-| ------------------------- | -------------------------                                                              | --------------------------- |
+| -------------------------- | -------------------------                                                              | --------------------------- |
 | DPH object                | $\tau\sim\text{DPH}(\boldsymbol{a}, \boldsymbol{T})$                                   | `DPH(T, a)`                 |
 | Mean                      | $\text{E}[\tau]=\boldsymbol{\pi} (\boldsymbol{I}-\boldsymbol{T})^{-1}\boldsymbol{e}$   | `mean(DPH)`                 |
 | Variance                  | $\text{V}[\tau]= \text{E}[\tau^2]-\text{E}[\tau]^2$                                    | `var(DPH)`                  |
 | Density                   | $f(x)=\boldsymbol{\pi T}^{\: x-1}\boldsymbol{t}$, $x\geq 1$                            | `dDPH(x, DPH)`              |
 | Cumulative distribution   | $F(x)=1-\boldsymbol{\pi T}^{\: x}\boldsymbol{e}$, $x\geq 1$                            | `pDPH(x, DPH)`              |
 | Quantile function         |                                                                                        | `qDPH(p, DPH)`              |
-| Random sampling.          |                                                                                        | `rDPH(n, DPH)`              |
+| Random sampling           |                                                                                        | `rDPH(n, DPH)`              |
 | Random sampling of full path         |                                                                             | `rFullDPH(n, DPH)`          |
 | Reward transformation     | $Y=\sum_{m=0}^{\tau-1} r(X_m)$                                                         | `reward_phase_type(DPH, r)` |
 
@@ -252,9 +252,9 @@ $$T_{\text{right}}=\min \bigl\{ t\geq 0 :X(t)\in\{5,6\} \bigr\}=\int_0^{\tau}\bo
 
 with the reward vector $\boldsymbol{r_{\text{right}}}=(1,1,1,1,0)$. A classical result in population genetics gives the covariance between the two tree heights
 
-$${\rm Cov}(T_{\text{left}},T_{\text{right}})=\frac{\rho+18}{\rho^2+13\rho+18},$$
+$$\text{Cov}(T_{\text{left}},T_{\text{right}})=\frac{\rho+18}{\rho^2+13\rho+18},$$
 
-and we note that for large recombination rates ${\rm Cov}(T_{\text{left}},T_{\text{right}})$ is close to zero, and for small recombination rates it is close to one. Also note that $T_{\text{left}}$ and $T_{\text{right}}$ are both exponentially distributed with a rate of 1, so $\text{Var}(T_{\text{left}})=\text{Var}(T_{\text{right}})=1$, and, consequently,  $\text{Cor}(T_{\text{left}}, T_{\text{right}})=\text{Cov}(T_{\text{left}}, T_{\text{right}})$ [see also equation 3.10 in @wakeley2009coalescent]. Moreover, as shown by a simple proof in @wilton2015smc, we have that $P(T_{\text{left}}=T_{\text{right}})=\text{Cov}(T_{\text{left}}, T_{\text{right}})$.
[…]
