### 2023-02-21 #5054 @Kevin-Mattheus-Moerman
@rivasiker I am the AEiC for this track and here to help process the final steps towards acceptance of this work in JOSS. Below are some minor points that need your attention.

On the archive:
- [ ] Ensure that the ZENODO listed license matches the software license, it currently says ` Other (Open)`, you may need to manually edit this.
- [ ] Ensure that the ZENODO title matches the paper title
- [ ] Ensure that the ZENODO author set matches those on the paper (including the order)

The paper seems in order. Note that you have no "Acknowledgements" section. This section is not required but if you do need it (e.g. to acknowledge funding) you can add one now if you like.
