### 2023-02-21 #5054 @rivasiker
@Kevin-Mattheus-Moerman, thank you for you pointing all of this out! I have updated the license information, title, and author list at the Zenodo archive, which can be found under the same DOI as before. Additionally,  I have added a small "Acknowledgements" section. With this, the draft would be finalized.
